Update CI and dev workflows (#1)
* Update CI and dev workflows

Add Makefile with shortcuts to make it easier to check and fix code
locally.

Update targeted python versions since 3.7 is no longer supported. The
chosen versions are the highest micro version of the highest and lowest
minor versions.

Add constraints file to improve reproducibility when bootstrapping
poetry. It is used with pip in CI because I can't be bothered figuring out
how to install packages with poetry such that they are available in
subsequent steps.

Add format-, lint- and type-checking since these are cheap ways to
improve code quality. Ruff is chosen because it is faster than pylint
and seems to have better defaults.

Import Self from typing extensions since it is not part of typing as of
python 3.8.

And probably more...