Configure rich handler to use stderr
... for consistency and to enable logging to be used also when the
output of a script is piped to another program.

Add a function that looks like a constructor in the handlers module to
create a consistent interface for selecting between named handlers and
to make it easier testing that this handler is compatible with other
options such as a custom format string.